Added functionalities to AprgCommon.

+++ AprgAudio/AprgAudio/tst/Audio/AprgAudio_unit.cpp
@@ -1,26 +1,24 @@
-#include <AprgAudio.hpp>
+#include <Audio/AprgAudio.hpp>
 
-#include <ExpectedValues/aiff_stereo_8bit_44100.hpp>
-#include <ExpectedValues/aiff_stereo_16bit_44100.hpp>
-#include <ExpectedValues/aiff_stereo_24bit_44100.hpp>
-#include <ExpectedValues/aiff_stereo_8bit_48000.hpp>
-#include <ExpectedValues/aiff_stereo_16bit_48000.hpp>
-#include <ExpectedValues/aiff_stereo_24bit_48000.hpp>
-#include <ExpectedValues/wav_mono_16bit_44100.hpp>
-#include <ExpectedValues/wav_mono_16bit_48000.hpp>
-#include <ExpectedValues/wav_stereo_8bit_44100.hpp>
-#include <ExpectedValues/wav_stereo_16bit_44100.hpp>
-#include <ExpectedValues/wav_stereo_24bit_44100.hpp>
-#include <ExpectedValues/wav_stereo_8bit_48000.hpp>
-#include <ExpectedValues/wav_stereo_16bit_48000.hpp>
-#include <ExpectedValues/wav_stereo_24bit_48000.hpp>
+#include <Audio/ExpectedValues/aiff_stereo_8bit_44100.hpp>
+#include <Audio/ExpectedValues/aiff_stereo_16bit_44100.hpp>
+#include <Audio/ExpectedValues/aiff_stereo_24bit_44100.hpp>
+#include <Audio/ExpectedValues/aiff_stereo_8bit_48000.hpp>
+#include <Audio/ExpectedValues/aiff_stereo_16bit_48000.hpp>
+#include <Audio/ExpectedValues/aiff_stereo_24bit_48000.hpp>
+#include <Audio/ExpectedValues/wav_mono_16bit_44100.hpp>
+#include <Audio/ExpectedValues/wav_mono_16bit_48000.hpp>
+#include <Audio/ExpectedValues/wav_stereo_8bit_44100.hpp>
+#include <Audio/ExpectedValues/wav_stereo_16bit_44100.hpp>
+#include <Audio/ExpectedValues/wav_stereo_24bit_44100.hpp>
+#include <Audio/ExpectedValues/wav_stereo_8bit_48000.hpp>
+#include <Audio/ExpectedValues/wav_stereo_16bit_48000.hpp>
+#include <Audio/ExpectedValues/wav_stereo_24bit_48000.hpp>
 #include <Math/AlbaMathHelper.hpp>
 
 #include <gtest/gtest.h>
-
 using namespace alba::mathHelper;
 using namespace std;
-
 namespace alba
 {
 
@@ -35,17 +33,15 @@ TEST(AiffLoadingTests, Stereo_8bit_44100)
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_8bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_8bit_44100::numChannels);
 
-    for (int i = 0; i < aiff_stereo_8bit_44100::testBuffer.size(); i++)
+    for (unsigned int i = 0; i < aiff_stereo_8bit_44100::testBuffer.size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_8bit_44100::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(AiffLoadingTests, Stereo_16bit_44100)
-{
+TEST(AiffLoadingTests, Stereo_16bit_44100){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\aiff_stereo_16bit_44100.aif)");
 
@@ -55,17 +51,15 @@ TEST(AiffLoadingTests, Stereo_16bit_44100)
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_16bit_44100::numChannels);
 
-    for (int i = 0; i < aiff_stereo_16bit_44100::testBuffer.size(); i++)
+    for (unsigned int i = 0; i < aiff_stereo_16bit_44100::testBuffer.size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_16bit_44100::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(AiffLoadingTests, Stereo_24bit_44100)
-{
+TEST(AiffLoadingTests, Stereo_24bit_44100){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\aiff_stereo_24bit_44100.aif)");
 
@@ -75,17 +69,15 @@ TEST(AiffLoadingTests, Stereo_24bit_44100)
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_24bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_24bit_44100::numChannels);
 
-    for (int i = 0; i < aiff_stereo_24bit_44100::testBuffer.size(); i++)
+    for (unsigned int i = 0; i < aiff_stereo_24bit_44100::testBuffer.size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_24bit_44100::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(AiffLoadingTests, Stereo_8bit_48000)
-{
+TEST(AiffLoadingTests, Stereo_8bit_48000){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\aiff_stereo_8bit_48000.aif)");
 
@@ -95,17 +87,15 @@ TEST(AiffLoadingTests, Stereo_8bit_48000)
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_8bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_8bit_48000::numChannels);
 
-    for (int i = 0; i < aiff_stereo_8bit_48000::testBuffer.size(); i++)
+    for (unsigned int i = 0; i < aiff_stereo_8bit_48000::testBuffer.size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_8bit_48000::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(AiffLoadingTests, Stereo_16bit_48000)
-{
+TEST(AiffLoadingTests, Stereo_16bit_48000){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\aiff_stereo_16bit_48000.aif)");
 
@@ -115,17 +105,15 @@ TEST(AiffLoadingTests, Stereo_16bit_48000)
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_16bit_48000::numChannels);
 
-    for (int i = 0; i < aiff_stereo_16bit_48000::testBuffer.size(); i++)
+    for (unsigned int i = 0; i < aiff_stereo_16bit_48000::testBuffer.size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_16bit_48000::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(AiffLoadingTests, Stereo_24bit_48000)
-{
+TEST(AiffLoadingTests, Stereo_24bit_48000){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\aiff_stereo_24bit_48000.aif)");
 
@@ -135,17 +123,15 @@ TEST(AiffLoadingTests, Stereo_24bit_48000)
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_24bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_24bit_48000::numChannels);
 
-    for (int i = 0; i < aiff_stereo_24bit_48000::testBuffer.size(); i++)
+    for (unsigned int i = 0; i < aiff_stereo_24bit_48000::testBuffer.size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_24bit_48000::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(WritingTest, WriteFromCopiedSampleBuffer)
-{
+TEST(WritingTest, WriteFromCopiedSampleBuffer){
     AprgAudio<float> audioFile1, audioFile2;
 
     bool isLoadingSuccessful = audioFile1.load (APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\wav_stereo_16bit_44100.wav)");
@@ -166,17 +152,15 @@ TEST(WavLoadingTests, Stereo_8bit_44100)
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_8bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_8bit_44100::testBuffer.size());
 
-    for (int i = 0; i < wav_stereo_8bit_44100::testBuffer[0].size(); i++)
+    for (unsigned int i = 0; i < wav_stereo_8bit_44100::testBuffer[0].size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_44100::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(WavLoadingTests, Stereo_16bit_44100)
-{
+TEST(WavLoadingTests, Stereo_16bit_44100){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\wav_stereo_16bit_44100.wav)");
 
@@ -186,17 +170,15 @@ TEST(WavLoadingTests, Stereo_16bit_44100)
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_16bit_44100::testBuffer.size());
 
-    for (int i = 0; i < wav_stereo_16bit_44100::testBuffer[0].size(); i++)
+    for (unsigned int i = 0; i < wav_stereo_16bit_44100::testBuffer[0].size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_44100::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(WavLoadingTests, Stereo_24bit_44100)
-{
+TEST(WavLoadingTests, Stereo_24bit_44100){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\wav_stereo_24bit_44100.wav)");
 
@@ -206,17 +188,15 @@ TEST(WavLoadingTests, Stereo_24bit_44100)
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_24bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_24bit_44100::testBuffer.size());
 
-    for (int i = 0; i < wav_stereo_24bit_44100::testBuffer[0].size(); i++)
+    for (unsigned int i = 0; i < wav_stereo_24bit_44100::testBuffer[0].size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_44100::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(WavLoadingTests, Mono_16bit_44100)
-{
+TEST(WavLoadingTests, Mono_16bit_44100){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\wav_mono_16bit_44100.wav)");
 
@@ -226,17 +206,15 @@ TEST(WavLoadingTests, Mono_16bit_44100)
     EXPECT_EQ(audioFile.getSampleRate(), wav_mono_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_mono_16bit_44100::numChannels);
 
-    for (int i = 0; i < wav_mono_16bit_44100::testBuffer.size(); i++)
+    for (unsigned int i = 0; i < wav_mono_16bit_44100::testBuffer.size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_mono_16bit_44100::testBuffer[i]));
-        }
-    }
+        }    }
 }
 
-TEST(WavLoadingTests, Stereo_8bit_48000)
-{
+TEST(WavLoadingTests, Stereo_8bit_48000){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\wav_stereo_8bit_48000.wav)");
 
@@ -246,17 +224,15 @@ TEST(WavLoadingTests, Stereo_8bit_48000)
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_8bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_8bit_48000::testBuffer.size());
 
-    for (int i = 0; i < wav_stereo_8bit_48000::testBuffer[0].size(); i++)
+    for (unsigned int i = 0; i < wav_stereo_8bit_48000::testBuffer[0].size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_48000::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(WavLoadingTests, Stereo_16bit_48000)
-{
+TEST(WavLoadingTests, Stereo_16bit_48000){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\wav_stereo_16bit_48000.wav)");
 
@@ -266,17 +242,15 @@ TEST(WavLoadingTests, Stereo_16bit_48000)
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_16bit_48000::testBuffer.size());
 
-    for (int i = 0; i < wav_stereo_16bit_48000::testBuffer[0].size(); i++)
+    for (unsigned int i = 0; i < wav_stereo_16bit_48000::testBuffer[0].size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_48000::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(WavLoadingTests, Stereo_24bit_48000)
-{
+TEST(WavLoadingTests, Stereo_24bit_48000){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\wav_stereo_24bit_48000.wav)");
 
@@ -286,17 +260,15 @@ TEST(WavLoadingTests, Stereo_24bit_48000)
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_24bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_24bit_48000::testBuffer.size());
 
-    for (int i = 0; i < wav_stereo_24bit_48000::testBuffer[0].size(); i++)
+    for (unsigned int i = 0; i < wav_stereo_24bit_48000::testBuffer[0].size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_48000::testBuffer[k][i]));
-        }
-    }
+        }    }
 }
 
-TEST(WavLoadingTests, Mono_16bit_48000)
-{
+TEST(WavLoadingTests, Mono_16bit_48000){
     AprgAudio<double> audioFile;
     bool isLoadingSuccessful = audioFile.load(APRG_AUDIO_DIR R"(\tst\ActualAudioFiles\wav_mono_16bit_48000.wav)");
 
@@ -306,34 +278,30 @@ TEST(WavLoadingTests, Mono_16bit_48000)
     EXPECT_EQ(audioFile.getSampleRate(), wav_mono_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_mono_16bit_48000::numChannels);
 
-    for (int i = 0; i < wav_mono_16bit_48000::testBuffer.size(); i++)
+    for (unsigned int i = 0; i < wav_mono_16bit_48000::testBuffer.size(); i++)
     {
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
         {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_mono_16bit_48000::testBuffer[i]));
-        }
-    }
+        }    }
 }
 
-bool writeTest(int numChannels, int sampleRate, int bitDepth, AprgAudioFormat format)
-{
+bool writeTest(int numChannels, int sampleRate, int bitDepth, AprgAudioFormat format){
     float sampleRateAsFloat = (float) sampleRate;
 
     AprgAudio<float> audioFile;
 
     audioFile.setAudioBufferSize (numChannels, sampleRate * 4);
 
-    for (int i = 0; i < audioFile.getNumSamplesPerChannel(); i++)
+    for (unsigned int i = 0; i < static_cast<unsigned int>(audioFile.getNumSamplesPerChannel()); i++)
     {
         float sample = sinf (2. * getPi() * ((float) i / sampleRateAsFloat) * 440.) ;
 
-        for (int k = 0; k < audioFile.getNumChannels(); k++)
+        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
             audioFile.samples[k][i] = sample * 0.5;
     }
-
     audioFile.setSampleRate (sampleRate);
     audioFile.setBitDepth (bitDepth);
-
     std::string numChannelsAsString = numChannels == 1 ? "mono" : "stereo";
     std::string bitDepthAsString = std::to_string (bitDepth);
     std::string sampleRateAsString = std::to_string (sampleRate);
